Extend this plugin to connect to BLE UART devices
This a major departure from the original point of this codebase. We will
now inject WebBLE code into any Meet windows with logic to connect back
to a BLE device advertising the Nordic Uart Service (NUS).

Then there's some application logic that's designed specifically for a
CircuitPython application to be running on a Nordic IC. When the mic is
muted/unmuted in meet, the extension will send commands to this
circuitPython application displaying status images on a screen. When the
cirucitPython application sends messages indicating buttons have been
pressed, the extension will mute and unmute the meet accordingly.